+++ lib/dart_frog_shared.dart
@@ -1,6 +1,10 @@
 /// Shared code for Dart Frog projects.
 library;
 
+export 'app_check/app_check_config.dart';
+export 'app_check/app_check_middleware.dart';
+export 'app_check/app_check_token_cache.dart';
+export 'app_check/firebase_app_check_service.dart';
 export 'exceptions/exceptions.dart';
 export 'exceptions/json_exportable.dart';
 export 'logging/log_handler.dart';